Compiling on Ubuntu 18.04. GLEW instead of GLAD

+++ include/nimble/opengl-wrapper/GLContext.h
@@ -1,4 +1,7 @@
 #pragma once
+
+#include <GL/glew.h>
+
 #include "nimble/utility/StrongTypes.h"
 
 // Singleton for global OpenGL handles/context/functions
@@ -14,6 +17,6 @@ public:
 	}
 
 	static void SetViewportDimensions(Width width, Height height) {
-		glViewport(0, 0, width.get(), height.get());
+		glViewport(0, 0, static_cast<int>(width.get()), static_cast<int>(height.get()));
 	}
 };
